requirements.txt
# 核心Web框架
fastapi==0.104.1
uvicorn==0.24.0

# 前端框架
streamlit

# HTTP客户端
aiohttp
requests

# 数据库和存储
chromadb

# 环境配置
python-dotenv

# 数据处理
pydantic

# 系统监控
psutil